feat(tests): move from multi to single opcode test
